requirements.txt
numpy
tqdm
pandas
matplotlib
python-chess
torch~=2.9.1
torchvision~=0.24.1
opencv-python


pillow~=12.0.0
pip~=24.2
filelock~=3.20.1
torchinfo